Fix forked SQLite connection handling to improve test accuracy (#3249)

+++ spec/beef/core/main/autorun_engine/autorun_engine_spec.rb
@@ -18,7 +18,10 @@ RSpec.describe 'AutoRunEngine Test', run_on_browserstack: true do
     print_info 'Loading database'
     db_file = @config.get('beef.database.file')
     print_info 'Resetting the database for BeEF.'
-    File.delete(db_file) if File.exist?(db_file)
+
+    if ENV['RESET_DB']
+      File.delete(db_file) if File.exist?(db_file)
+    end
 
     @config.set('beef.credentials.user', 'beef')
     @config.set('beef.credentials.passwd', 'beef')
@@ -49,9 +52,13 @@ RSpec.describe 'AutoRunEngine Test', run_on_browserstack: true do
       OTR::ActiveRecord.establish_connection!
     end
     ActiveRecord::Migrator.migrations_paths = [File.join('core', 'main', 'ar-migrations')]
-    context = ActiveRecord::MigrationContext.new(ActiveRecord::Migrator.migrations_paths)
-    ActiveRecord::Migrator.new(:up, context.migrations, context.schema_migration, context.internal_metadata).migrate if context.needs_migration?
-
+    MUTEX.synchronize do
+      context = ActiveRecord::MigrationContext.new(ActiveRecord::Migrator.migrations_paths)
+      if context.needs_migration?
+        ActiveRecord::Migrator.new(:up, context.migrations, context.schema_migration, context.internal_metadata).migrate
+      end
+    end
+    
     BeEF::Core::Migration.instance.update_db!
 
     # add AutoRunEngine rule
@@ -63,16 +70,17 @@ RSpec.describe 'AutoRunEngine Test', run_on_browserstack: true do
     # Spawn HTTP Server
     print_info 'Starting HTTP Hook Server'
     http_hook_server = BeEF::Core::Server.instance
-    http_hook_server.prepare
 
     # Generate a token for the server to respond with
     @token = BeEF::Core::Crypto.api_token
 
+    # ***** IMPORTANT: close any and all AR/OTR connections before forking *****
+    disconnect_all_active_record!
+
     # Initiate server start-up
-    @pids = fork do
-      BeEF::API::Registrar.instance.fire(BeEF::API::Server, 'pre_http_start', http_hook_server)
-    end
     @pid = fork do
+      http_hook_server.prepare
+      BeEF::API::Registrar.instance.fire(BeEF::API::Server, 'pre_http_start', http_hook_server)
       http_hook_server.start
     end
 
@@ -101,6 +109,7 @@ RSpec.describe 'AutoRunEngine Test', run_on_browserstack: true do
 
   after(:all) do
     server_teardown(@driver, @pid, @pids)
+    disconnect_all_active_record!
   end
 
   it 'AutoRunEngine is working' do
